chatgptwebapp/requirements.txt
Django
langchain==0.0.278
numexpr==2.8.5
openai==0.28.1
whitenoise
wsgi-basic-auth
tiktoken
opencensus
opencensus-ext-azure
opencensus-ext-django
django-sequences
psycopg2
pgvector
django-rest-swagger
djangorestframework
pdfminer
azure-storage-blob
python-dotenv
numpy==1.21.2
pandas==1.3.2
matplotlib==3.4.2
japanize-matplotlib==1.1.3
tabulate==0.8.9
openpyxl==3.0.7
xlsxwriter==1.4.3
faker==8.15.0